Разнес 4 класса по разным файлам, удалил лишний файл, упростил чуть чуть код. В общем исправил все замечания.
