--- app/src/main/java/com/example/playlistmaker/SearchMysicVievHolder.kt
@@ -1,17 +1,0 @@
-package com.example.playlistmaker
-
-import android.view.LayoutInflater
-import android.view.RoundedCorner
-import android.view.View
-import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-
-
-class SearchMysicVievHolder {
-
-}
-
