+++ app/src/main/java/com/example/playlistmaker/Track.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker
+
+data class Track(
+    val trackName: String, // Название композиции
+    val artistName: String, // Имя исполнителя
+    val trackTime: String, // Продолжительность трека
+    val artworkUrl100: String, // Ссылка на изображение обложки
+)
